+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(DCC)
+
+test_check("DCC")